pyproject.toml
[tool.poetry]
name = "ewtwitterbot"
version = "0.1.0"
description = "A stupid Twitter bot that responds to requests with Explorers Wanted quotes or markov sentences."
authors = ["Daniel Andrlik <daniel@andrlik.org>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
tweepy = {extras = ["async"], version = "^4.5.0"}
Pillow = "^9.0.1"
loguru = "^0.6.0"
"Mastodon.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
types-requests = "^2.27.10"
black = {extras = ["d"], version = "^22.1.0"}
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
pytest-sugar = "^0.9.4"
requests-mock = "^1.9.3"
types-Pillow = "^9.0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
